requirements.txt
streamlit==1.31.1
google-generativeai==0.4.0
python-dotenv==1.0.1
ipython==8.22.2